feat!: listen for AuthorizationEvent to notify users of permission request
When authorization is required, the Client.AuthorizationRequired event is
dispatched as an instance of Client.AuthorizationEvent at the document level,
with an authorization_url property. The user should be notified of the need
for the permission request, with suitable explanation, and then redirected
to the authorizaton_url (ideally in a new window).

BREAKING: Authorization is no longer abstracted away. The package consumer
MUST direct the user to the authorization_url, ideally with suitable
explanation.